third_party/blink/web_tests/platform/generic/fast/events/zoom-dblclick-expected.txt




Zoomed
PASS event.clientX is 166
PASS event.clientY is 83
PASS event.pageX is 166
PASS event.pageY is 83
PASS successfullyParsed is true

TEST COMPLETE
